requirements.txt
flask==3.1.3
numpy==2.4.3
pandas==3.0.1
scikit-learn==1.8.0
xgboost==3.2.0
requests==2.33.1
python-dotenv==1.2.2
groq==1.1.2
kagglehub==1.0.0
pymongo==4.10.1
